Make curr_func not keep history
Previously curr_func was a stack of all the previous function names.
This would allow for compiled recursive functions. Unfortunately, its hard to know
when a function in a .vm file is truly over, because there can be multiple
return calls. So, we are just keeping the topmost current function.